rf hot fix for rows_sample

+++ cpp/src/randomforest/randomforest.cu
@@ -500,7 +500,7 @@ RF_metrics cross_validate(const cumlHandle& user_handle,
 RF_params set_rf_class_obj(int max_depth, int max_leaves, float max_features,
                            int n_bins, int split_algo, int min_rows_per_node,
                            bool bootstrap_features, bool bootstrap, int n_trees,
-                           int rows_sample) {
+                           float rows_sample) {
   DecisionTree::DecisionTreeParams tree_params(
     max_depth, max_leaves, max_features, n_bins, split_algo, min_rows_per_node,
     bootstrap_features);
